Rewrite / add documentation
Instead of the controller and view sharing the same API but different methods,
all the methods are now kept on the controller level and the view simply defers
to them. The main problem with keeping them separate was that it was possible
for @_window_titles to be different between the controller and the view. Now
there's only one copy of that array floating around.